feat: new method of broadCast update
