+++ backend/src/websocket/handlers/joinRoom.handler.ts
@@ -1,5 +1,8 @@
 import { ConnectedUser, activeConnections } from "../index";
-import { roomManager } from "../managers/roomManager";
+import {
+    roomManager,
+    broadcastRoomListUpdateToLobby,
+} from "../managers/roomManager";
 
 export function handleJoinRoom(
     user: ConnectedUser,
@@ -62,19 +65,5 @@ export function handleJoinRoom(
     };
     roomManager.broadcastToRoom(room.id, JSON.stringify(roomInfoPayload));
 
-    broadcastRoomListUpdate();
-}
-
-function broadcastRoomListUpdate() {
-    const publicRooms = roomManager.getPublicRooms();
-    const roomListPayload = {
-        event: "room_list",
-        rooms: publicRooms,
-    };
-
-    for (const connection of activeConnections) {
-        if (connection.roomId === null) {
-            connection.ws.send(JSON.stringify(roomListPayload));
-        }
-    }
+    broadcastRoomListUpdateToLobby(activeConnections);
 }
