+++ backend/src/websocket/handlers/leaveRoom.handler.ts
@@ -1,6 +1,9 @@
 import { ConnectedUser, activeConnections } from "../index";
 
-import { roomManager } from "../managers/roomManager";
+import {
+    roomManager,
+    broadcastRoomListUpdateToLobby,
+} from "../managers/roomManager";
 
 export function handleLeaveRoom(user: ConnectedUser) {
     const { roomId } = user;
@@ -40,19 +43,5 @@ export function handleLeaveRoom(user: ConnectedUser) {
         roomManager.removeRoom(room.id);
     }
 
-    broadcastRoomListUpdate();
-}
-
-function broadcastRoomListUpdate() {
-    const publicRooms = roomManager.getPublicRooms();
-    const roomListPayload = {
-        event: "room_list",
-        rooms: publicRooms,
-    };
-
-    for (const connection of activeConnections) {
-        if (connection.roomId === null) {
-            connection.ws.send(JSON.stringify(roomListPayload));
-        }
-    }
+    broadcastRoomListUpdateToLobby(activeConnections);
 }
